add/update: 機能ごとに関数切り出し
